Using SessionActor instead of MainActor
